requirements.txt
langchain==0.1.0
openai==1.7.0
chromadb==0.4.24
python-dotenv==1.0.1
langchain_openai==0.0.2
langchain-community==0.0.11
pymongo==4.6.3
fastapi==0.110.1
uvicorn==0.29.0
gunicorn==22.0.0
tqdm
streamlit
streamlit-chat
